Removed server store/action with reaction + removed alt.js
