--- app/scripts/alt.jsx
@@ -1,3 +1,0 @@
-import Alt from 'alt';
-
-export default new Alt();
