+++ app/scripts/components/channel.jsx
@@ -1,21 +1,10 @@
 import React from 'react';
-import connectToStores from 'alt/utils/connectToStores';
-
-import ServerStore from '../stores/servers';
 
 import UserList from './userlist';
 import Chat from './chat';
 import Input from './input';
 
-@connectToStores
 export default class ChannelView extends React.Component {
-    static getStores() { return [ServerStore]; }
-    static getPropsFromStores() {
-        return {
-            servers: ServerStore.getState().servers
-        };
-    }
-
     shouldComponentUpdate(newProps) {
         const oldServerId = this.props.params.serverId;
         const newServerId = newProps.params.serverId;
@@ -26,8 +15,8 @@ export default class ChannelView extends React.Component {
         const oldMessages = this.props.state.messages[oldServerId].channels[oldChannelName];
         const newMessages = newProps.state.messages[newServerId].channels[newChannelName];
 
-        const oldChannel = this.props.servers[oldServerId].channels[oldChannelName];
-        const newChannel = newProps.servers[newServerId].channels[newChannelName];
+        const oldChannel = this.props.state.servers[oldServerId].channels[oldChannelName];
+        const newChannel = newProps.state.servers[newServerId].channels[newChannelName];
 
         return oldMessages !== newMessages || oldChannel !== newChannel;
     }
