+++ app/scripts/mainApp.jsx
@@ -8,6 +8,7 @@ import State from './stores/state';
 import './reactions/config';
 import './reactions/route';
 import './reactions/message';
+import './reactions/server';
 
 import SquelchView from './components/squelchView';
 import WelcomeView from './components/welcome';
