+++ package.json
@@ -34,7 +34,6 @@
     "rimraf": "^2.4.1"
   },
   "dependencies": {
-    "alt": "^0.17.1",
     "babel-core": "^5.6.15",
     "bootstrap": "^3.3.5",
     "classnames": "^2.1.3",
